split up concerns even more into individual methods
