+++ app/controllers/concerns/redirect_if_error.rb
@@ -1,0 +1,14 @@
+module RedirectIfError
+  extend ActiveSupport::Concern
+
+  included do
+    def redirect_if_error(object)
+      unless object.save
+        flash[:error] = object.errors.full_messages.join(', ')
+        redirect_to action: :edit
+      end
+    end
+  end
+end
+
+Comfy::Admin::Cms::PagesController.send(:include, RedirectIfError)
