+++ config/initializers/comfortable_mexican_sofa.rb
@@ -123,6 +123,12 @@ ComfortableMexicanSofa::AccessControl::AdminAuthentication.password = ENV["COMFY
 #   end
 # end
 
-require File.expand_path("../../../app/models/concerns/icca_links.rb", __FILE__)
-require File.expand_path("../../../app/models/concerns/fragment_search.rb", __FILE__)
-require File.expand_path("../../../app/controllers/concerns/photos_links_resources.rb", __FILE__)
+# CONCERNS
+require File.expand_path("../../../app/models/concerns/has_many_custom.rb", __FILE__)
+require File.expand_path("../../../app/models/concerns/link_icca_site.rb", __FILE__)
+require File.expand_path("../../../app/models/concerns/link_country.rb", __FILE__)
+require File.expand_path("../../../app/models/concerns/search_integration.rb", __FILE__)
+require File.expand_path("../../../app/controllers/concerns/redirect_if_error.rb", __FILE__)
+require File.expand_path("../../../app/controllers/concerns/save_photos.rb", __FILE__)
+require File.expand_path("../../../app/controllers/concerns/save_links.rb", __FILE__)
+require File.expand_path("../../../app/controllers/concerns/save_resources.rb", __FILE__)
